clean up some gcc 4 64-bit compiler warnings
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1013 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/lib/expat/xmlparse/xmlparse.c
@@ -2693,8 +2693,8 @@ processContentToken(XML_Parser       const xmlParserP,
     case XML_TOK_INVALID:
         *eventPP = *nextP;
         *errorCodeP = XML_ERROR_INVALID_TOKEN;
-        xmlrpc_asprintf(errorP, "Invalid token, starting %u bytes in",
-                        *nextP - s);
+        xmlrpc_asprintf(errorP, "Invalid token, starting %ld bytes in",
+                        (long)(*nextP - s));
         break;
     case XML_TOK_PARTIAL:
         if (nextPtr) {
